Update README.md
Goal of Whole Genome Sequencing (WGS)

The primary goal of Whole Genome Sequencing (WGS) is to determine the complete DNA sequence of an organism's genome at a single time. This comprehensive approach aims to identify the genetic variations and mutations that may be responsible for diseases, traits, and evolutionary processes. By providing a detailed map of an organism's genetic makeup, WGS facilitates advanced research in genetics, personalized medicine, and evolutionary biology.